requirements.txt
discord.py[voice]>=2.3.2
python-dotenv>=1.0.0
yt-dlp>=2023.10.13
PyNaCl>=1.5.0 